requirements.txt
fastapi
uvicorn
pytesseract
pillow
python-docx
pymupdf
spacy
transformers
textblob
python-dotenv
